androidclient: Model amount type is now float while unit is now WeightUnits

+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/api/mapper/output/OutputIngredientMapper.kt
@@ -8,7 +8,7 @@ class OutputIngredientMapper {
 
     fun mapToOutputModel(model: MealIngredient) = IngredientOutput(
         identifier = requireNotNull(model.submissionId),
-        quantity = model.amount
+        quantity = model.amount.toInt()
     )
 
     fun mapToOutputModelCollection(models: Collection<MealIngredient>) =
