+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/db/entity/DbPortionEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = DbPortionEntity.tableName)
 open class DbPortionEntity(
-    val amount: Int
+    val amount: Float
 ) {
     companion object {
         const val DEFAULT_DB_ID: Long = 0
